refactor: split massive 2700-line orb.test.ts into focused test files
- Split single monolithic test file into 12 function-specific test files
- Improved maintainability and navigation
- Reduced merge conflicts and cognitive load
- All 146 tests continue to pass with 90.22% coverage maintained
- Each test file focuses on a single function with consistent structure

Files created:
- create-customer.test.ts (5 tests)
- create-subscription.test.ts (6 tests)
- get-customer-subscriptions.test.ts (4 tests)
- get-customer-details.test.ts (3 tests)
- schedule-plan-change.test.ts (8 tests)
- unschedule-plan-change.test.ts (3 tests)
- remove-fixed-fee-quantity-transition.test.ts (3 tests)
- edit-fixed-fee-quantity-transitions.test.ts (4 tests)
- add-price-interval.test.ts (6 tests)
- edit-price-interval-schedule.test.ts (8 tests)
- get-price-details.test.ts (8 tests)
